requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-image>=0.21.0
tqdm>=4.65.0
tensorboard>=2.14.0
PyYAML>=6.0
einops>=0.7.0
timm>=0.9.7
matplotlib>=3.7.0
pandas>=2.0.0
rasterio>=1.3.0
albumentations>=1.3.1
lpips>=0.1.4
